refactor: split data into modular files
