+++ lib/data/index.ts
@@ -1,0 +1,4 @@
+export * from "./links";
+export * from "./experiences";
+export * from "./projects";
+export * from "./skills";
