+++ lib/data/links.ts
@@ -1,0 +1,8 @@
+export const links = [
+  { name: "Home", hash: "#home" },
+  { name: "About", hash: "#about" },
+  { name: "Experience", hash: "#experience" },
+  { name: "Projects", hash: "#projects" },
+  { name: "Skills", hash: "#skills" },
+  { name: "Contact", hash: "#contact" },
+] as const;
