+++ lib/data/skills.ts
@@ -1,0 +1,35 @@
+export const skillsData = [
+	"Python",
+	"Java",
+	"JavaScript",
+  "C++",
+	"R",
+	"SQL",
+	"Airflow",
+	"Spark",
+	"Hadoop",
+	"Tableau",
+	"Looker Studio",
+	"Power BI",
+	"PyTorch",
+	"TensorFlow",
+  "Scikit-learn",
+	"LlamaIndex",
+	"Langchain",
+	"MongoDB",
+	"Django",
+	"Firebase",
+	"HTML",
+	"CSS",
+	"React",
+	"Next.js",
+	"Vue.js",
+	"Redux",
+  "Tailwind",
+	"Material UI",
+	"Bootstrap",
+  "Git",
+	"Excel",
+	"Alteryx",
+] as const;
+
